refactor(tests): consolidate diffusion payload classes into shared payloads.py
